fixed login issue

+++ routes/web.php
@@ -20,24 +20,25 @@ use App\Http\Controllers\UserController;
 Route::get('/', [ListingController::class, 'index'] );
 
 
+//->middleware('auth'); will make user be authed
 
 //show create form
-Route::get('/listings/create', [ListingController::class, 'create'] );
+Route::get('/listings/create', [ListingController::class, 'create'] )->middleware('auth');
 
 
-Route::post('/listings', [ListingController::class, 'store'] );
+Route::post('/listings', [ListingController::class, 'store'] )->middleware('auth');
 
 //single listing
 Route::get('/listings/{listing}',[ListingController::class, 'show'] );
 
 //edit
-Route::get('/listings/{listing}/edit',[ListingController::class, 'edit'] );
+Route::get('/listings/{listing}/edit',[ListingController::class, 'edit'] )->middleware('auth');;
 
 //update
-Route::put('/listings/{listing}',[ListingController::class, 'update'] );
+Route::put('/listings/{listing}',[ListingController::class, 'update'])->middleware('auth');;
 
 //delete
-Route::delete('/listings/{listing}',[ListingController::class, 'destroy'] );
+Route::delete('/listings/{listing}',[ListingController::class, 'destroy'] )->middleware('auth');;
 
 
 //users
@@ -47,9 +48,9 @@ Route::get('/register',[UserController::class, 'create'] );
 Route::post('/users',[UserController::class, 'store'] );
 
 //logout
-Route::post('/logout',[UserController::class, 'logout'] );
+Route::post('/logout',[UserController::class, 'logout'] )->middleware('auth');
 
-Route::get('/login',[UserController::class, 'login'] );
+Route::get('/login',[UserController::class, 'login'] )->name('login');
 
 Route::post('/users/authenticate',[UserController::class, 'authenticate'] );
